Finish migration to boto3 (instead of s3fs)

+++ lakota/s3_pod.py
@@ -1,10 +1,15 @@
 from pathlib import Path
 
 import boto3
+import urllib3
 from botocore.exceptions import ClientError
 
 from .pod import POD
 from .utils import logger
+
+
+def silence_insecure_warning():
+    urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 
 class S3POD(POD):
@@ -22,26 +27,30 @@ class S3POD(POD):
         secret=None,
         token=None,
     ):
+        """
+        `path` must contains bucket name and sub-path in the bucket
+        (separated by a `/`). `verify` set to False will disable ssl
+        verifications. Please note that boto3 will also use
+        "REQUESTS_CA_BUNDLE" env variable.
+        """
         bucket, *parts = path.parts
         self.path = Path(*parts)
         self.bucket = Path(bucket)
         if client:
             self.client = client
         else:
-            # Disable ssl verifications if asked. Please note that
-            # boto will take "REQUESTS_CA_BUNDLE" env variable.
-            client_kwargs = {
-                "use_ssl": True if verify else None,
-            }
-            if netloc:
-                # TODO support for https on custom endpoints
-                # TODO document use of param: endpoint_url='http://127.0.0.1:5300'
-                client_kwargs["endpoint_url"] = f"http://{netloc}"
-            self.client = boto3.client(
-                "s3",
+            # TODO support for https on custom endpoints
+            # TODO document use of param: endpoint_url='http://127.0.0.1:5300'
+            if not verify:
+                silence_insecure_warning()
+            endpoint_url = f"http://{netloc}" if netloc else None
+            session = boto3.session.Session(
                 aws_access_key_id=key,
                 aws_secret_access_key=secret,
-                # **client_kwargs, #FIXME
+                profile_name=profile,
+            )
+            self.client = session.client(
+                "s3", verify=verify, aws_session_token=token, endpoint_url=endpoint_url
             )
         super().__init__()
 
@@ -52,8 +61,9 @@ class S3POD(POD):
     def ls(self, relpath=".", missing_ok=False, limit=None):
         logger.debug("LIST s3:///%s/%s %s", self.bucket, self.path, relpath)
         paginator = self.client.get_paginator("list_objects")
-        prefix = str(self.path / relpath) + "/"
-        cut = len(str(self.path)) + 1
+        prefix = str(self.path / relpath)
+        prefix = "" if prefix in (".", "") else prefix + "/"
+        cut = len(prefix)
         options = {
             "Bucket": str(self.bucket),
             "Prefix": prefix,
@@ -65,16 +75,23 @@ class S3POD(POD):
         page_iterator = paginator.paginate(**options)
         names = []
         for page in page_iterator:
+            # Extract pseudo-folder names
             common_prefixes = page.get("CommonPrefixes", [])
+            names.extend(item["Prefix"][cut:].rstrip("/") for item in common_prefixes)
+            # Extract keys (filenames)
             contents = page.get("Contents", [])
-            names.extend(item["Prefix"][cut:] for item in common_prefixes)
             names.extend(item["Key"][cut:] for item in contents)
         return names
 
     def read(self, relpath, mode="rb"):
         logger.debug("READ s3:///%s/%s %s", self.bucket, self.path, relpath)
         key = str(self.path / relpath)
-        resp = self.client.get_object(Bucket=str(self.bucket), Key=key)
+        try:
+            resp = self.client.get_object(Bucket=str(self.bucket), Key=key)
+        except ClientError as err:
+            if err.response["Error"]["Code"] == "NoSuchKey":
+                raise FileNotFoundError(f'Key "{relpath}" not found')
+
         return resp["Body"].read()
 
     def write(self, relpath, data, mode="wb"):
@@ -107,12 +124,34 @@ class S3POD(POD):
 
     def rm(self, relpath=".", recursive=False, missing_ok=False):
         logger.debug("REMOVE s3://%s/%s %s", self.bucket, self.path, relpath)
-        key = str(self.path / relpath)
-        _ = self.client.delete_object(
-            Bucket=str(self.bucket),
-            Key=key,
-        )
-        # TODO implement missing_ok
+        prefix = str(self.path / relpath)
+        if recursive:
+            prefix = "" if prefix in (".", "") else prefix + "/"
+            resp = self.client.list_objects_v2(Bucket=str(self.bucket), Prefix=prefix)
+            keys = [item["Key"] for item in resp.get("Contents", [])]
+        else:
+            keys = [prefix]
+        try:
+            _ = self.client.delete_objects(
+                Bucket=str(self.bucket),
+                Delete={
+                    "Objects": [{"Key": k} for k in keys],
+                    "Quiet": True,
+                },
+            )
+            # TODO check for error in response
+        except ClientError as err:
+            if err.response["Error"]["Code"] != "MalformedXML":
+                raise
+            # As of version 2.2.6, Moto doesn't support correctly
+            # delete_objects() calls, we fall back to delete_object()
+            for key in keys:
+                self.client.delete_object(
+                    Bucket=str(self.bucket),
+                    Key=key,
+                )
+
+        # XXX implement missing_ok
 
     def mv(self, from_path, to_path, missing_ok=False):
         orig = str(self.path / from_path)
@@ -121,10 +160,16 @@ class S3POD(POD):
             "MOVE s3://%s/%s to s3://%s/%s", self.bucket, orig, self.bucket, dest
         )
         try:
+            # Copy to dest
             self.client.copy(
                 {"Bucket": str(self.bucket), "Key": orig},
                 str(self.bucket),
                 dest,
+            )
+            # Delete orig
+            self.client.delete_object(
+                Bucket=str(self.bucket),
+                Key=orig,
             )
         except ClientError as err:
             if err.response["Error"]["Code"] == "404":
